bug fixes and path search ui changes

+++ app/src/main/java/com/archidni/archidni/Model/Path/MoveInstruction.java
@@ -10,9 +10,10 @@ import java.util.ArrayList;
 
 public abstract class MoveInstruction extends PathInstruction {
 
-    public MoveInstruction(int duration) {
+    public MoveInstruction(long duration) {
         super(duration);
     }
+
     public abstract float getDistance ();
     public abstract ArrayList<Coordinate> getPolyline ();
 }
